created basic login page and router for home page
