+++ src/App.js
@@ -1,15 +1,33 @@
-import Login from './Login';
-import SignUp from './SignUp';
+import Login from './pages/Login';
+import SignUp from './pages/SignUp';
 import AuthDetails from './AuthDetails';
-import './App.css'
+import Home from './pages/Home';
+import './App.css';
+import {BrowserRouter as Router, Route, Switch} from 'react-router-dom'
 
 function App() {
   return (
-    <div className="App">
-      <Login/>
-      <SignUp/>
-      <AuthDetails/>
+    <Router>
+      <div className="App">
+        <div className='content'>
+          <Switch>
+            <Route exact path='/'>
+              <Login/>
+              <SignUp/>
+              <AuthDetails/>
+            </Route>
+            <Route path='/home'>
+              <Home/>
+            </Route>
+          </Switch>
+
+        </div>
+
      </div>
+
+
+    </Router>
+   
   );
 }
