+++ src/index.js
@@ -7,8 +7,7 @@ import reportWebVitals from './reportWebVitals';
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    /<App />
-
+    <App />
   </React.StrictMode>
 );
