+++ src/navigation/NavBar.js
@@ -1,5 +1,6 @@
 import React, {useState} from "react"
-import logo from './UMKC_Logo.png';
+import logo from '../UMKC_Logo.png';
+import {Link} from 'react-router-dom';
 
 function NavBar() {
     const [open, setOpen] = useState(false);
@@ -9,7 +10,9 @@ function NavBar() {
             <nav>
                 <div className="logo"> <img src={logo} alt="UMKC" /> </div>
                 <ul className="nav-links" style={{transform: open ? "translateX(0px)" : "" }}>
-                    <li><a>Home</a></li>
+                    <Link to='/home'>
+                        <li><a>Home</a></li>
+                    </Link>
                     <li><a>My Account</a></li>
                     <li><a>Submission Portal</a></li>
                     {/*<li><a>About</a></li>*/}
